io_utils split because of cross importing
