+++ src/python/shared/BAData.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass, field
 import numpy as np
 
 from shared.BASparseMat import BASparseMat
-from shared.io_utils import save_errors_to_file, objective_file_name,\
+from shared.output_utils import save_errors_to_file, objective_file_name,\
                             save_sparse_j_to_file, jacobian_file_name
