+++ src/python/shared/HandData.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass, field
 from typing import Tuple
 import numpy as np
 
-from shared.io_utils import save_vector_to_file, objective_file_name,\
+from shared.output_utils import save_vector_to_file, objective_file_name,\
                             save_jacobian_to_file, jacobian_file_name
 
 @dataclass
